Local bugfix: continue search in a handler cluster after decline
Optimization of HANDLER-BIND, %HANDLER-BIND (with handler clusters
represented as dx closures using typecase for dispatching) introduced
a bug: when multiple typespecs match and the first handler declines,
no further search within the same handler cluster (that is, the same
handler-bind form) was done. I was taking "visual analogy with
typecase" [CLHS handler-bind] too far.

Make it a chain of singular TYPECASEs instead of an all-encompassing
typecase, so it behaves properly in this respect.